--- a/src/Render/Camera.h
+++ b/src/Render/Camera.h
@@ -73,16 +73,6 @@ public:
     }
 
     void Rasterize(Scene* scene) {
-        int numTriangles = 0;
-    
-        //for each object in the scene, get the triangles W
-        for(int i = 0; i < scene->GetObjectCount(); i++){
-            if(scene->GetObjects()[i]->IsEnabled()){
-                numTriangles += scene->GetObjects()[i]->GetTriangleGroup()->GetTriangleCount();
-            }
-        }
-        
-        Triangle2D** triangles = new Triangle2D*[numTriangles];
         int triangleCounter = 0;
         
         lookDirection = transform->GetRotation().Conjugate() * lookOffset;
@@ -101,12 +91,7 @@ public:
             if(scene->GetObjects()[i]->IsEnabled()){
                 //for each triangle in object, project onto 2d surface, but pass material
                 for (int j = 0; j < scene->GetObjects()[i]->GetTriangleGroup()->GetTriangleCount(); j++) {
-                    triangles[triangleCounter] = new Triangle2D(lookDirection, transform, &scene->GetObjects()[i]->GetTriangleGroup()->GetTriangles()[j], scene->GetObjects()[i]->GetMaterial());
-
-                    bool triangleInView = tree.insert(triangles[triangleCounter]);
-                    
-                    if(triangleInView) triangleCounter++;
-                    else delete triangles[triangleCounter];//out of view space remove from array
+                    tree.insert(Triangle2D(lookDirection, transform, &scene->GetObjects()[i]->GetTriangleGroup()->GetTriangles()[j], scene->GetObjects()[i]->GetMaterial()));
                 }
             }
         }
@@ -123,12 +108,6 @@ public:
 
             pixelGroup->GetPixel(i)->Color = CheckRasterPixel(leafNode->entities, leafNode->count, pixelRay);
         }
-
-        for (int i = 0; i < triangleCounter; i++){
-            delete triangles[i];
-        }
-        
-        delete[] triangles;
     }
     
 };