--- a/src/Render/Scene.h
+++ b/src/Render/Scene.h
@@ -20,7 +20,6 @@ public:
 	}
 
 	~Scene(){
-        printf("destroying scene\n");
 		delete[] objects;
 	}