This PR provides a preview of a proposed screenspace quadtree implementation for accelerating ray-triangle intersection testing. The rendering loop is modified to include the following steps:
- determine the screenspace bounds
- initialize a quadtree to the screen bounds
- insert each projected triangle into the tree
- for each pixel, determine the intersecting leaf node and test against the contained triangles

initial benchmarks show a significant performance increase for scenes with moderate to large triangle counts, some initial results can be found below.

| Scene | Platform | Render Time (original) | Render Time (quadtree) | Speedup|
|--------|-----------|-------------------------|----------------------------|----------|
|SpyroAnimation|x86|3.82ms|0.45ms| 8.49x|

Note that this implementation is not yet in a final stage, however the current state can already be used for testing and performance evaluation. For a final release the following TODOs will be addressed:
- reduce memory usage by avoiding unnecessary duplicate references to triangles
- adapt root node bounding box to intersection of screen bounds and triangle bounds
- evaluate performance on more platforms and scenes