@coelacant1 2022-09-27
Thank you @moepforfreedom, just merged it in with some slight modifications. I separated the classes and decoupled them /added mutators for the variable calls. This added a significant performance boost, great work!