@moepforfreedom 2022-09-28
im glad i could help. there are probably a few more ways to increase the overall rendering performance so i might submit a few more PRs in the near future.